# app/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "EpiLog API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/users": {
      "get": {
        "summary": "Alle Nutzer abrufen",
        "tags": [
          "Users"
        ],
        "responses": {
          "200": {
            "description": "Liste aller Nutzer"
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "500": {
            "description": "Serverfehler"
          }
        }
      },
      "post": {
        "summary": "Nutzer anlegen",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "username",
                  "password",
                  "firstname"
                ],
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "firstname": {
                    "type": "string"
                  },
                  "lastname": {
                    "type": "string"
                  },
                  "quote": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Nutzer erfolgreich angelegt"
          },
          "400": {
            "description": "Pflichtfelder fehlen"
          },
          "500": {
            "description": "Serverfehler"
          }
        }
      },
      "patch": {
        "summary": "Eigenes Profil aktualisieren",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstname": {
                    "type": "string"
                  },
                  "lastname": {
                    "type": "string"
                  },
                  "quote": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profil erfolgreich aktualisiert"
          },
          "401": {
            "description": "Nicht autorisiert"
          },
          "500": {
            "description": "Serverfehler"
          }
        }
      }
    }
  },
...
}